flowgrind: daily commit - various minor and major bugfixes, unified DEBUG_MSG
git-svn-id: svn+ssh://svn.umic-mesh.net/umic-mesh/projects/flowgrind/branches/flowgrind-next@6359 3d4b872d-e00e-0410-a2ce-a5c2cda4747c

+++ src/daemon.h
@@ -91,10 +91,10 @@ struct _flow
 		long long bytes_read;
 		long long bytes_written;
 
-		long request_blocks_read;
-		long request_blocks_written;
-		long response_blocks_read;
-		long response_blocks_written;
+		int request_blocks_read;
+		int request_blocks_written;
+		int response_blocks_read;
+		int response_blocks_written;
 
 		double iat_min, iat_max, iat_sum;
 		double rtt_min, rtt_max, rtt_sum;
